Don't depend on socket state changes for signals
In the current implementation of QAbstractSocket, when an error occurs
that causes a state change the state change signal is sometimes emitted
before the error signal, making watching for the Disconnected state an
unreliable way to fire `finished()` since that should only happen once
everything with the interface has concluded (including all errors being
communicated).

Instead watch for `disconnected()`, and `errorOccurred()` to fire the
signal as those both cover all cases where the sockets state reverts to
'Disconnected".