Revert non state-change approach (problematic)
This approach is problematic since there are many ways that a socket
could "end" that need to be accounted for, in a much less clean way that
using QAbstractSocket::stateChanged(), including being aborted,
disconnecting, having a socket error (but only when a connection was
never established), etc.

Overall this is likely more error prone than the approach that will
follow this revert.

NOTE: If these changes are ever to be re-used, be aware that the abort
case of needing to triggered finished() has not yet been handled.